Implement basic typing options
The typing experience only worked one way. Now, it responds to four settings. They're not perfect, but they're mostly there:
  - allow-wrong-input: print wrong characters. Let the user backspace, as is natural.
  - require-capitalization: toggles whether the user can type everything in lowercase.
  - require-punctuation: toggles the ability to auto-insert/auto-type succeeding punctuation characters.
  - require-newline: toggles the ability to auto-insert newlines where appropriate.

There are potential design issues with this implementation. For example, the apostrophe in "you're" probably should not be printed after typing "u" because that's a possibly undesired hint, and this is a memorization app.

Resolve #5

Unrelated: minor convention change ($varname for jQuery objects).