Accept equivalent characters
When pasting content, some characters might be very hard to type, such as emdashes or curly quotes. The equivalent characters (dashes and straight quotes) should be accepted instead, and should produce the intended character rather than the input character (so that the produced output looks identical to the source text).

A dictionary of equivalencies was created. Willing to explore alternative solutions. Maybe a library out there deals with this?