Implement textarea size and auto-scroll
Size: The "input textarea" (should come up with a better name) now varies in size according to the amount of text inside it: a minimum of 3 lines tall, a maximum of about 50% of the window and everything inbetween to shrink-wrap the text inside it.
Auto-scroll: Like the input textarea, the output box is often too short for certain texts. For better usability, in cases where the typed text overflows, the output box will quickly but gently scroll down so that the user can always see the characters they're typing. This is far better than manual scrolling, and better than having the text immediately jump.