Use build tools
Now using Browserify and other tools to create/copy "distribution" files to a "dist" root directory (not versioned).
Using npm scripts as the "task runner" instead of unwieldy alternatives like Gulp or Grunt.
Also changed the one library to be used as a node module rather than an object in `window`