Moar build scripts
Fine-tuning these scripts, which, considering my level of ignorance and inexperience, means significant reengineering.

There is only one script to be used for production: "build" which puts everything in dist/. The other scripts are for dev, but should generally only be run via the "start" script.

Further re-engineering lies ahead: uglify.js, stylus compilation, and html transforms (to inject the livereload tag, for example, or the application version number).