feat(semver): new utility for parsing app versions
This tiny little utility will allow the application to know when a new version of the app is loaded, and to apply release-specific transforms to any existing saved texts. So, if version n lacks some data, the app can, at startup, check the last app version and, if lesser than the current app version, can run a one-time function on existing texts to update them with the necessary data, or with appropriate placeholders, so that the main application code can be free from making such checks and assume a single data definition.