feat(semver): update user data pending app version
On page load, if the new app is newer than the app the last time it ran, update
functions will run that will progressively update existing data. This
encapsulates the problem of, say, textObj properties not being set, and so
eliminates the need for a lot of sanitation out of the functions that
operate on textObj.

Each function will only ever run once, if ever. This means that the application
may bloat over time. Ideally, the textObj interface will not change much after
the first major release, so the extra bytes should stay manageable.
Additionally, we can eliminate a lot of the 0.x.y update functions before the
1.0.0 milestone is reached.