feat(textsCrud, modal): allow user to set initial review for new texts
Instead of having to start from the very beginning every time, an experienced memorizer can add well-known texts to the application to use its SRS to manage a large backlog of reviews, but without having to go through the hassle of checking YES today, then tomorrow, then in two days, etc. Visual feedback tells the user when the next due date will be, both date-wise and as a number of days from today.